requirements.txt
fastapi
python-dotenv
uvicorn[standard]
pydantic
celery
langchain
openai
tiktoken
yt_dlp
scenedetect[opencv]
Pillow
faiss-cpu
sentence-transformers
numpy
multilingual-clip
open_clip_torch
pytube
git+https://github.com/m-bain/whisperx.git